Replace challenges/mixins by utils and add 'proposals' property to each XXX-proposal components
