+++ live/app/components/qcu-proposals.js
@@ -1,5 +1,6 @@
 import Ember from 'ember';
 import labeledCheckboxes from 'pix-live/utils/labeled-checkboxes';
+import proposalsAsArray from 'pix-live/utils/proposals-as-array';
 
 function _uncheckAllRadioButtons() {
   this.$(':radio').prop('checked', false);
@@ -11,10 +12,9 @@ function _checkAgainTheSelectedOption(index) {
 
 export default Ember.Component.extend({
 
-  tagName: 'div',
-
   labeledRadios: Ember.computed('proposals', 'answers', function() {
-    return labeledCheckboxes(this.get('proposals'), this.get('answers'));
+    const arrayOfProposals = proposalsAsArray(this.get('proposals'));
+    return labeledCheckboxes(arrayOfProposals, this.get('answers'));
   }),
 
   actions: {
