+++ live/app/models/challenge.js
@@ -1,11 +1,9 @@
 import Ember from 'ember';
 import DS from 'ember-data';
-import ProposalsAsArrayMixin from './challenge/proposals-as-array-mixin';
-import ProposalsAsBlocksMixin from './challenge/proposals-as-blocks-mixin';
 
 const { Model, attr } = DS;
 
-export default Model.extend(ProposalsAsArrayMixin, ProposalsAsBlocksMixin, {
+export default Model.extend({
 
   instruction: attr('string'),
   proposals: attr('string'),
