+++ live/app/templates/components/challenge-item-qroc.hbs
@@ -15,7 +15,7 @@
   <div class="rounded-panel challenge-response">
     <div class="rounded-panel__row challenge-proposals">
       {{qroc-proposal
-        blocks=challenge._proposalsAsBlocks
+        proposals=challenge.proposals
         answerValue=answers.value
         onInputChanged='inputChanged'}}
     </div>
