+++ live/app/templates/components/challenge-item-qrocm.hbs
@@ -15,7 +15,7 @@
   <div class="rounded-panel challenge-response">
     <div class="rounded-panel__row challenge-proposals">
       {{qrocm-proposal
-        blocks=challenge._proposalsAsBlocks
+        proposals=challenge.proposals
         answersValue=answers._valuesAsMap
         onInputChanged='inputChanged' }}
     </div>
