+++ live/app/templates/components/qroc-proposal.hbs
@@ -1,4 +1,4 @@
-{{#each blocks as |block|}}
+{{#each _blocks as |block|}}
 
   {{#if block.text}}
     <span>{{block.text}}</span>
