+++ live/app/templates/components/qrocm-proposal.hbs
@@ -1,4 +1,4 @@
-{{#each blocks as |block|}}
+{{#each _blocks as |block|}}
 
   {{#if block.text}}
     <span>{{block.text}}</span>
@@ -13,7 +13,7 @@
   {{/if}}
 
   {{#if block.breakline}}
-    <br> 
+    <br>
   {{/if}}
 
 {{/each}}
