+++ live/tests/integration/components/qroc-proposal-test.js
@@ -1,37 +1,34 @@
-import Ember from 'ember';
 import {expect} from 'chai';
 import {describe, it} from 'mocha';
 import {setupComponentTest} from 'ember-mocha';
 import hbs from 'htmlbars-inline-precompile';
 
-describe('Integration | Component | QrocProposal', function() {
+describe('Integration | Component | QROC proposal', function() {
 
   setupComponentTest('qroc-proposal', {
     integration: true
   });
 
-  beforeEach(function() {
-    const block = [];
-    block.push(Ember.Object.create({name: 'myInput', input: 'mylabel'}));
-    this.set('blocks', block);
+  it('renders', function() {
+    this.render(hbs`{{qroc-proposal}}`);
+    expect(this.$()).to.have.length(1);
   });
 
   describe('Component behavior when user fill input of challenge:', function() {
 
-    it('renders', function() {
-      this.render(hbs`{{qroc-proposal}}`);
-      expect(this.$()).to.have.length(1);
-    });
-
     it('should display a value when a non-empty value is providing by user', function() {
       // given
+      const proposals = '${myInput}';
+      this.set('proposals', proposals);
       this.set('answerValue', 'myValue');
       // when
-      this.render(hbs`{{qroc-proposal blocks=blocks answerValue=answerValue}}`);
+      this.render(hbs`{{qroc-proposal proposals=proposals answerValue=answerValue}}`);
       // then
-      expect(this.$('.challenge-response__proposal-input').val()).to.be.equal('myValue');
+      expect(this.$('.challenge-response__proposal-input').val()).to.equal('myValue');
     });
   });
+
+  //     block.push(Ember.Object.create({name: 'myInput', input: 'mylabel'}));
 
   describe('Component behavior when user skip challenge:', function() {
 
@@ -48,9 +45,10 @@ describe('Integration | Component | QrocProposal', function() {
 
       it(`should display '' value ${input} is providing to component`, function() {
         // given
+        this.set('proposals', '${myLabel}');
         this.set('answerValue', input);
         // when
-        this.render(hbs`{{qroc-proposal blocks=blocks answerValue=answerValue}}`);
+        this.render(hbs`{{qroc-proposal proposals=proposals answerValue=answerValue}}`);
         // then
         expect(this.$('.challenge-response__proposal-input').val()).to.be.equal(output);
       });
